Add default functions in BenchmarkModule trait
This refactoring reduces a bit the number of lines of LoC by deduplicating the empty functions in the trait directly
