+++ src/benchmark/empty.rs
@@ -1,6 +1,4 @@
-use crate::benchmark::default::IntervalCounter;
-use crate::benchmark::{BenchmarkModule, Counter, Scope};
-use crate::virt::VirtContext;
+use crate::benchmark::BenchmarkModule;
 
 pub struct EmptyBenchmark {}
 
@@ -11,27 +9,5 @@ impl BenchmarkModule for EmptyBenchmark {
 
     fn name() -> &'static str {
         "Empty Benchmark"
-    }
-
-    fn start_interval_counters(_scope: Scope) {}
-
-    fn stop_interval_counters(_scope: Scope) {}
-
-    fn increment_counter(_counter: Counter) {}
-
-    fn update_inteval_counter_stats(
-        &mut self,
-        _counter: &IntervalCounter,
-        _scope: &Scope,
-        _value: usize,
-    ) {
-    }
-
-    fn read_counters(_ctx: &mut VirtContext) {}
-
-    fn display_counters() {}
-
-    fn get_counter_value(_core_id: usize, _counter: Counter) -> usize {
-        0
     }
 }
